refactor: replace bincode serialization with parity_scale_codec

+++ config/src/lib.rs
@@ -4,9 +4,9 @@ pub static CONFIG: LazyLock<RwLock<Config>> = LazyLock::new(|| RwLock::new(load_
 
 #[derive(Debug)]
 pub struct Config {
-    pub single_tx_max_size: usize,
-    pub tx_max_size: usize,
-    pub min_tx_threshold: usize,
+    pub single_tx_max_size: u64,
+    pub tx_max_size: u64,
+    pub min_tx_threshold: u64,
 }
 
 pub fn load_config() -> Config {
