+++ transaction/src/lib.rs
@@ -45,12 +45,12 @@ mod tests {
         let test_single_tx_size = (config.tx_max_size - config.min_tx_threshold) / test_tx_count;
 
         for _ in 0..test_tx_count {
-            pool.add_tx(Transaction::dummy_size(test_single_tx_size))
+            pool.add_tx(Transaction::dummy_size(test_single_tx_size as usize))
                 .await
                 .unwrap();
         }
 
-        pool.add_tx(Transaction::dummy_size(test_tx_count + 1))
+        pool.add_tx(Transaction::dummy_size(test_tx_count as usize + 1))
             .await
             .unwrap();
